Separate client from cluster
