+++ pkg/cluster/k3d.go
@@ -11,11 +11,7 @@ import (
 	"github.com/rancher/k3d/v5/pkg/config/v1alpha3"
 	"github.com/rancher/k3d/v5/pkg/runtimes"
 	"github.com/rancher/k3d/v5/pkg/types"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 )
 
 const (
@@ -25,11 +21,6 @@ const (
 
 var agentArgs = []string{
 	"--node-label=nebula.puppet.com/scheduling.customer-ready=true",
-}
-
-type Client struct {
-	APIClient client.Client
-	Mapper    meta.RESTMapper
 }
 
 // K3dClusterManager wraps rancher's k3d to manage the lifecycle
@@ -177,39 +168,6 @@ func (m *K3dClusterManager) WriteKubeconfig(ctx context.Context, path string) er
 	}
 
 	return k3dclient.KubeconfigWriteToPath(ctx, apiConfig, path)
-}
-
-// GetClient returns a new Client configured with a RESTMapper and k8s api client.
-func (m *K3dClusterManager) GetClient(ctx context.Context, opts ClientOptions) (*Client, error) {
-	apiConfig, err := m.GetKubeconfig(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	overrides := &clientcmd.ConfigOverrides{}
-	clientConfig := clientcmd.NewDefaultClientConfig(*apiConfig, overrides)
-
-	restConfig, err := clientConfig.ClientConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	c, err := client.New(restConfig, client.Options{
-		Scheme: opts.Scheme,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	mapper, err := apiutil.NewDynamicRESTMapper(restConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	return &Client{
-		APIClient: c,
-		Mapper:    mapper,
-	}, nil
 }
 
 func (m *K3dClusterManager) get(ctx context.Context) (*types.Cluster, error) {
